backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
# asyncpg: needed for Supabase PostgreSQL in production (install separately: pip install asyncpg)
# asyncpg==0.30.0

# Supabase
supabase==2.10.0

# Redis (optional for development)
redis==5.2.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.27,<0.28

# Socket.IO
python-socketio==5.12.1

# AI Integration
openai==1.58.1

# Payments
razorpay==1.4.2

# Environment
python-dotenv==1.0.1
pydantic-settings==2.7.0

# Image Processing
Pillow>=10.0.0

# PDF Generation
reportlab==4.2.5

# File Storage
minio>=7.2.0

# Utilities
python-dateutil==2.9.0
uuid==1.30
